# specs/001-family-library-mvp/contracts/api.yaml
openapi: 3.0.3
info:
  title: Family Library API
  description: Corporate family management system for FreeAxez Revit projects
  version: 1.0.0
  contact:
    name: FreeAxez BIM Team
servers:
  - url: http://localhost:5000/api
    description: Local development
  - url: https://family-library.freeaxez.com/api
    description: Production
...
paths:
...
  /families:
    get:
      tags: [Families]
      summary: Search families
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: tagId
          in: query
          schema:
            type: string
            format: uuid
        - name: roleId
          in: query
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Paginated list of families
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedFamilies'
  /families/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Families]
      summary: Get family details
      responses:
        '200':
          description: Family details with versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyDetail'
        '404':
          description: Family not found
  /families/{id}/versions:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Families]
      summary: Get family version history
      responses:
        '200':
          description: List of versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FamilyVersion'
...
  /families/publish:
    post:
      tags: [Families]
      summary: Publish family to library
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                rfaFile:
                  type: string
                  format: binary
                txtFile:
                  type: string
                  format: binary
                metadata:
                  type: object
                  properties:
                    roleId:
                      type: string
                      format: uuid
                    familyName:
                      type: string
                    contentHash:
                      type: string
                    commitMessage:
                      type: string
      responses:
        '201':
          description: Family published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyVersion'
  /families/validate-hash:
    post:
      tags: [Families]
      summary: Check for hash duplicates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contentHash:
                  type: string
                excludeFamilyId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Duplicate check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuplicateCheckResult'
  /families/batch-check:
    post:
      tags: [Families]
      summary: Batch check family statuses
      description: Check multiple families for updates available
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                families:
                  type: array
                  items:
                    type: object
                    properties:
                      roleName:
                        type: string
                      contentHash:
                        type: string
      responses:
        '200':
          description: Status for each family
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FamilyStatus'

  # ========== SYSTEM TYPES ==========
...
components:
  schemas:
...
    PagedResponse:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer

    # ========== ROLES ==========
...
    Family:
      type: object
      properties:
        id:
          type: string
          format: uuid
        roleId:
          type: string
          format: uuid
        roleName:
          type: string
        familyName:
          type: string
        currentVersion:
          type: integer
        hasCatalog:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FamilyDetail:
      allOf:
        - $ref: '#/components/schemas/Family'
        - type: object
          properties:
            versions:
              type: array
              items:
                $ref: '#/components/schemas/FamilyVersion'
            types:
              type: array
              items:
                $ref: '#/components/schemas/FamilyType'
    FamilyVersion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        version:
          type: integer
        hash:
          type: string
        previousHash:
          type: string
        originalFileName:
          type: string
        originalCatalogName:
          type: string
        commitMessage:
          type: string
        publishedAt:
          type: string
          format: date-time
        publishedBy:
          type: string
    FamilyType:
      type: object
      properties:
        name:
          type: string
        parameters:
          type: object
          additionalProperties:
            type: string
    PagedFamilies:
      allOf:
        - $ref: '#/components/schemas/PagedResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Family'
    DuplicateCheckResult:
      type: object
      properties:
        isDuplicate:
          type: boolean
        existingFamily:
          $ref: '#/components/schemas/Family'
    FamilyStatus:
      type: object
      properties:
        roleName:
          type: string
        contentHash:
          type: string
        status:
          type: string
          enum: [UpToDate, UpdateAvailable, LegacyMatch, NotInLibrary]
        libraryVersion:
          type: integer
          nullable: true

    # ========== SYSTEM TYPES ==========
...